Add changes to lab 1

+++ Labs/Lab1/src/main/java/mk/finki/ukim/mk/lab/model/EventBooking.java
@@ -9,10 +9,10 @@ public class EventBooking {
     String attendeeAddress;
     Long numberOfTickets;
 
-    public EventBooking(String eventName, String attendeeName, String attendeeAddress, int numberOfTickets) {
+    public EventBooking(String eventName, String attendeeName, String attendeeAddress, Long numberOfTickets) {
         this.eventName = eventName;
         this.attendeeName = attendeeName;
         this.attendeeAddress = attendeeAddress;
-        this.numberOfTickets = (long) numberOfTickets;
+        this.numberOfTickets = numberOfTickets;
     }
 }
